backend/requirements.txt
flask
flask-cors
torch
torchvision
pillow
diffusers
transformers
accelerate
safetensors
PyJWT
bcrypt
